Complete adding bistim info to real-time apps
In real-time apps:

Bistim (both stimulators values and ISI) is sent to the recruitment curve and SICI axes.

The effective stimulator value is computed and saved in the datapoints table/csv file, even though it is not used in the RC figure.

The stimulation mode (bistim or simultaneous_discharge) is saved in rc_fit_info.txt file.

Bug  with the activity monitor axes being zoomed by scrolling on the mouse has been fixed: set axes Interactions to [].