Adding more bistim info to review_emg.rc.mlapp
Changes for review_emg_rc.mlapp:

Write the stimulator hardware setup to the datapoints.csv file.

Edit rc_boltzman_fit.m to use the effective stimulator output and save the stimulator_mode to the info struct so it will be saved to file.

Add display of multiple emg data lines in the axes and their average.