update init_rc_fig.m so rc figure displays in correct position on Brainsight
rc figure window position is specified for the layout of windows I like on brainsight. On other computers this may be off the screen. So use movegui.m to move it onto the screen for other computers. If movegui is used on brainsight, then the window is moved to be about a half inch in from the edge of the screen - not where I want it.